Added press and hold feature for one button.

+++ Keyboard.cpp
@@ -1,6 +1,6 @@
 #include "pch.h"
 #include "Keyboard.h"
-#include "Hooks.h" // Needed for the original function pointers (fp...)
+#include "Hooks.h"
 
 namespace ScreenshotInput {
 
@@ -11,26 +11,26 @@ namespace ScreenshotInput {
 
     SHORT WINAPI Keyboard::HookedGetAsyncKeyState(int vKey) {
         if (samekeyA == vKey) {
-            return 0x8001;
+            return (short)0x8001;
         }
         samekeyA = 0;
 
         if (vKey == keystatesend) {
             samekeyA = vKey;
-            return 0x8000;
+            return (short)0x8000;
         }
         return Hooks::fpGetAsyncKeyState(vKey);
     }
 
     SHORT WINAPI Keyboard::HookedGetKeyState(int nVirtKey) {
         if (samekey == nVirtKey) {
-            return 0x8001;
+            return (short)0x8001;
         }
         samekey = 0;
-
+        
         if (nVirtKey == keystatesend) {
             samekey = nVirtKey;
-            return 0x8000;
+            return (short)0x8000;
         }
         return Hooks::fpGetKeyState(nVirtKey);
     }
